requirements.txt

kafka-python==2.0.2
psutil==5.9.5
docker==6.1.3
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0.1
click==8.1.7
tabulate==0.9.0
colorama==0.4.6
tqdm==4.66.1

